adding files

+++ Scripts/01-Sample_processing_and_selection/Functions.sh
@@ -28,30 +28,32 @@ task_Fastqc1(){
 		echo -e "\nExecute fastqc..."
 		fastqc $1"_1.fastq.gz"
 		fastqc $1"_2.fastq.gz"
-		## Save fastqc results in fastqc folder.
-		echo -e "\nMove fastqc files to fastqc folder..."
-		mv $1"_1_fastqc.zip" ./fastqc
-		mv $1"_1_fastqc.html" ./fastqc
-		mv $1"_2_fastqc.zip" ./fastqc
-		mv $1"_2_fastqc.html" ./fastqc
-	else
+		## Save fastqc results in Fastqc folder.
+		echo -e "\nMove fastqc files to Fastqc folder..."
+		mv $1"_1_fastqc.zip" ./Fastqc
+		mv $1"_1_fastqc.html" ./Fastqc
+		mv $1"_2_fastqc.zip" ./Fastqc
+		mv $1"_2_fastqc.html" ./Fastqc
+	elif [ -e $1".fastq" ]; then
 		## Compress fastq files to fastq.gz.
 		echo -e "\nCompress the library..."
 		pigz -p $2 $1".fastq"
 		## Execute fastqc.
 		echo -e "\nExecute fastqc..."
 		fastqc $1".fastq.gz"
-		## Save fastqc results in fastqc folder.
-		echo -e "\nMove fastqc files to fastqc folder..."
-		mv $1"_fastqc.zip" ./fastqc
-		mv $1"_fastqc.html" ./fastqc
+		## Save fastqc results in Fastqc folder.
+		echo -e "\nMove fastqc files to Fastqc folder..."
+		mv $1"_fastqc.zip" ./Fastqc
+		mv $1"_fastqc.html" ./Fastqc
+	else
+		echo -e "\n$1 doesn't exist. Check it."
 	fi
 }
 
 task_Trimming(){
 	echo -e "\n"$1"..."
 	
-	if [ -e ../01-Raw_data/$1"_2.fastq.gz" ]; then
+	if [ -e ../01-Raw_data/$1"_1.fastq.gz" ] && [ -e ../01-Raw_data/$1"_2.fastq.gz" ]; then
 		## Trimming the libraries.
 		echo -e "\nTrimming..."
 		fastp \
@@ -74,8 +76,7 @@ task_Trimming(){
 			--length_required 49 \
 			--trim_poly_x \
 			--poly_x_min_len 10
-		
-	else
+	elif [ -e ../01-Raw_data/$1".fastq.gz" ]; then
 		## Trimming the libraries.
 		echo -e "\nTrimming..."
 		fastp \
@@ -95,6 +96,8 @@ task_Trimming(){
 			--length_required 49 \
 			--trim_poly_x \
 			--poly_x_min_len 10
+	else
+		echo -e "\n$1 doesn't exist. Check it."
 	fi
 }
 
@@ -106,20 +109,22 @@ task_Fastqc2(){
 		echo -e "\nExecute fastqc..."
 		fastqc $1"_tr_1P.fastq.gz"
 		fastqc $1"_tr_2P.fastq.gz"
-		## Save fastqc results in fastqc folder.
-		echo -e "\nMove fastqc files to fastqc folder..."
-		mv $1"_tr_1P_fastqc.zip" ./fastqc
-		mv $1"_tr_1P_fastqc.html" ./fastqc
-		mv $1"_tr_2P_fastqc.zip" ./fastqc
-		mv $1"_tr_2P_fastqc.html" ./fastqc
-	else
+		## Save fastqc results in Fastqc folder.
+		echo -e "\nMove fastqc files to Fastqc folder..."
+		mv $1"_tr_1P_fastqc.zip" ./Fastqc
+		mv $1"_tr_1P_fastqc.html" ./Fastqc
+		mv $1"_tr_2P_fastqc.zip" ./Fastqc
+		mv $1"_tr_2P_fastqc.html" ./Fastqc
+	elif [ -e $1"_tr.fastq.gz" ]; then
 		## Execute fastqc.
 		echo -e "\nExecute fastqc..."
 		fastqc $1"_tr.fastq.gz"
-		## Save fastqc results in fastqc folder.
-		echo -e "\nMove fastqc files to fastqc folder..."
-		mv $1"_tr_fastqc.zip" ./fastqc
-		mv $1"_tr_fastqc.html" ./fastqc
+		## Save fastqc results in Fastqc folder.
+		echo -e "\nMove fastqc files to Fastqc folder..."
+		mv $1"_tr_fastqc.zip" ./Fastqc
+		mv $1"_tr_fastqc.html" ./Fastqc
+	else
+		echo -e "\n$1 doesn't exist. Check it."
 	fi
 }
 
@@ -135,62 +140,77 @@ task_Summary_table_trimming(){
 		num_sequences_2=$(($total_lines_2/4))
 		## Check if the libraries are empty files.
 		echo -e "\nCheck if the libraries are empty files..."
-		if [ "$(zcat $1_tr_1P.fastq.gz | wc -l)" -gt 0 ] && [ "$(zcat $1_tr_2P.fastq.gz | wc -l)" -gt 0 ]; then
+		if [ $num_sequences_1 -ge 1000000 ] && [ $num_sequences_2 -ge 1000000 ]; then
 			echo -e $1" passes the trimmomatic step."
-			echo -e $1"\tpasses the trimmomatic step\tYES\t$num_sequences_1/$num_sequences_2" >> $2
-			echo $1 >> $3
-		elif [ "$(zcat $1_tr_1P.fastq.gz | wc -l)" -gt 0 ] || [ "$(zcat $1_tr_2P.fastq.gz | wc -l)" -gt 0 ]; then
+			echo -e $1"\tpasses the trimmomatic step\tYES\t$num_sequences_1/$num_sequences_2" >> $2/$1.tsv
+			echo $1 >> $3/$1.txt
+		elif [ $num_sequences_1 -ge 1000000 ] || [ $num_sequences_2 -ge 1000000 ]; then
 			echo -e $1" passes the filter only forward or reverse)."
-			echo -e $1"\tpasses the filter only forward or reverse\tNO\t$num_sequences_1/$num_sequences_2" >> $2
+			echo -e $1"\tpasses the filter only forward or reverse\tNO\t$num_sequences_1/$num_sequences_2" >> $2/$1.tsv
 		else
 			echo -e $1" doesn't pass the trimmomatic step."
-			echo -e $1"\tdoesn't pass the trimmomatic step\tNO\t$num_sequences_1/$num_sequences_2" >> $2
+			echo -e $1"\tdoesn't pass the trimmomatic step\tNO\t$num_sequences_1/$num_sequences_2" >> $2/$1.tsv
 		fi
-	else
+	elif [ -e $1"_tr.fastq.gz" ]; then
 		## Calculate the library depth.
 		echo -e "\nCalculate the library depth..."
 		total_lines=$(zcat $1"_tr.fastq.gz" | wc -l)
 		num_sequences=$(($total_lines/4))
 		## Check if the library is an empty file.
 		echo -e "\nCheck if the library is an empty file..."
-		if [ "$(zcat $1_tr.fastq.gz | wc -l)" -gt 0 ]; then
+		if [ $num_sequences -ge 1000000 ]; then
 			echo -e $1" passes the trimmomatic step."
-			echo -e $1"\tpasses the trimmomatic step\tYES\t$num_sequences" >> $2
-			echo $1 >> $3
+			echo -e $1"\tpasses the trimmomatic step\tYES\t$num_sequences" >> $2/$1.tsv
+			echo $1 >> $3/$1.txt
 		else
 			echo -e $1" doesn't pass the trimmomatic step."
-			echo -e $1"\tdoesn't pass the trimmomatic step\tNO\t$num_sequences" >> $2
+			echo -e $1"\tdoesn't pass the trimmomatic step\tNO\t$num_sequences" >> $2/$1.tsv
 		fi
+	else
+		echo -e "\n$1 doesn't exist. Check it."
 	fi
 }
 
 task_Pseudomapping(){
 	echo -e "\n"$1"..."
 	
-	if [ -e $2/02-Trimmed_data/$1"_tr_2P.fastq.gz" ]; then
+	if [ -e $2/02-Trimmed_data/$1"_tr_1P.fastq.gz" ] && [ -e $2/02-Trimmed_data/$1"_tr_2P.fastq.gz" ]; then
 		echo -e "\nPseudomapping (Libraries vs reference transcriptome)..."
-		salmon quant -i $2/03-Strand_detection/02-Index/$3 -l A \
-		-1 $2/02-Trimmed_data/$1"_tr_1P.fastq.gz" -2 $2/02-Trimmed_data/$1"_tr_2P.fastq.gz" \
+		salmon quant \
+		-i $2/03-Strand_detection/02-Index/$3 \
+		-l A \
+		-1 $2/02-Trimmed_data/$1"_tr_1P.fastq.gz" \
+		-2 $2/02-Trimmed_data/$1"_tr_2P.fastq.gz" \
 		-g $4/GFF3_genes/$3.gff3 \
-		-p $5 -o $1 --gcBias
-	else
+		-p $5 \
+		-o $1 \
+		--gcBias
+	elif [ -e $2/02-Trimmed_data/$1"_tr.fastq.gz" ]; then
 		echo -e "\nPseudomapping (Library vs reference transcriptome)..."
-		salmon quant -i $2/03-Strand_detection/02-Index/$3 -l A \
+		salmon quant \
+		-i $2/03-Strand_detection/02-Index/$3 \
+		-l A \
 	    	-r $2/02-Trimmed_data/$1"_tr.fastq.gz" \
 	    	-g $4/GFF3_genes/$3.gff3 \
-	    	-p $5 -o $1 --gcBias
+	    	-p $5 \
+	    	-o $1 \
+	    	--gcBias
+	else
+		echo -e "\n$1 doesn't exist. Check it."
 	fi
 }
 
 task_Select_SS_libraries(){
 	
-	if [ -e $2/02-Trimmed_data/$1"_tr_2P.fastq.gz" ]; then
-		echo -e "\nMove the strand-specific libraries to 04-Selected_data..."
+	if [ -e $2/02-Trimmed_data/$1"_tr_1P.fastq.gz" ] && [ -e $2/02-Trimmed_data/$1"_tr_2P.fastq.gz" ]; then
+		echo -e "Move the strand-specific library $1 to 04-Selected_data..."
 		cp $2/02-Trimmed_data/$1"_tr_1P.fastq.gz" ./
 		cp $2/02-Trimmed_data/$1"_tr_2P.fastq.gz" ./
-	else
-		echo -e "\nMove the strand-specific library to 04-Selected_data..."
+	elif [ -e $2/02-Trimmed_data/$1"_tr.fastq.gz" ]; then
+		echo -e "Move the strand-specific library $1 to 04-Selected_data..."
 		cp $2/02-Trimmed_data/$1"_tr.fastq.gz" ./
+	else
+		echo -e "\n$1 doesn't exist. Check it."
 	fi
 }
